app-health-back/src/@api/app-health/application/resolvers/app-health-get-applications.resolver.spec.ts
/* eslint-disable @typescript-eslint/no-unused-vars */
import { Test, TestingModule } from '@nestjs/testing';

// custom items
import { AppHealthGetApplicationsResolver } from './app-health-get-applications.resolver';
import { AppHealthGetApplicationsHandler } from '../handlers/app-health-get-applications.handler';

// sources
import { appHealthMockApplicationData } from '@app/app-health/application/infrastructure/mock/app-health-mock-application.data';

describe('AppHealthGetApplicationsResolver', () =>
{
    let resolver: AppHealthGetApplicationsResolver;
    let handler: AppHealthGetApplicationsHandler;

    beforeAll(async () =>
    {
        const module: TestingModule = await Test.createTestingModule({
            imports: [
            ],
            providers: [
                AppHealthGetApplicationsResolver,
                {
                    provide : AppHealthGetApplicationsHandler,
                    useValue: {
                        main: () => { /**/ },
                    },
                },
            ],
        })
            .compile();

        resolver = module.get<AppHealthGetApplicationsResolver>(AppHealthGetApplicationsResolver);
        handler = module.get<AppHealthGetApplicationsHandler>(AppHealthGetApplicationsHandler);
    });

    test('AppHealthGetApplicationsResolver should be defined', () =>
    {
        expect(resolver).   toBeDefined();
    });

    describe('main', () =>
    {
        test('AppHealthGetApplicationsResolver should be defined', () =>
        {
            expect(resolver).   toBeDefined();
        });

        test('should return a appHealthMockApplicationData', async () =>
        {
            jest.spyOn(handler, 'main').mockImplementation(() => new Promise(resolve => resolve(appHealthMockApplicationData)));
            expect(await resolver.main()).toBe(appHealthMockApplicationData);
        });
    });
});